move the file to the module directory
