+++ modules/functional-util/index.js
